requirements.txt
flask
validators
shortuuid
pytest
pytest-mock
flask-sqlalchemy
flask-cors
python-dotenv